Update release Dockerfile to contain labels
For the relases we discussed in https://github.com/elastic/package-registry/issues/375 that we want to have the revisions for debugging purpose. This adds both and adjust the Dockerfile to rely on git clone instead of go get.

Additional change is to adjust the not logging requests for /healthcheck as otherwise this gets spammed in the logs.