Cleanup example packages
Thanks to the Beats module migration script, most of the packages which were built as example packages are now available in a more perfect format inside the Beats packages. Because of this most of the existing example packages were removed except for the ones with long example logs or the reference package.

The only two left are system and aws packages as these still have some special configs inside, but as soon as the PRs to get a never version of these in are merged, they will also be removed.

The endpoint and base package were moved to the alpha directory as these are the packages intended to be released for the alpha release.