Add support for datastreams on base templates
The v2 templates are updated to use datastreams. This means now moving forward all data is ingested using datastreams. This requires at least Elasticesarch 7.9.

The data_stream config is put into the template itself and not one of the components as this part should never change.

In this PR, the base package is readded to the repository as currently no other place exists for development and testing.