Remove requirement for streams definition in dataset manifest (#483)
Previously streams definition in the dataset manifest was required. This led to the problem that for example the endpoint package specified an input there even though it is not used: https://github.com/elastic/package-storage/blob/master/packages/endpoint/0.1.0/dataset/events/manifest.yml This removes the requirement for the streams definition and adds a package for testing it.