Add /types API endpoint
The type API endpoint exposes the number of packages with a dataset of a certain type exists. Today this is what is exposed by the category endpoint but this will be repurposed. The value of the type is based on the dataset.type field, indirectly the type definition inside the dataset manifest.

It is expected that Kibana switches over to using this endpoint and then the category endpoint can be refactored. What is currently exposed under category is not really categories.