Add list of downloads to /search endpoint
As discussed in https://github.com/elastic/package-registry/issues/474 in the future we might have more then one format to download a package. To be prepared for this, downloads becomes an array.

In this PR, the existing `download` entry is not removed yet. It will be removed as soon as Kibana switched over the usage.