Rename `kibana` query param to `kibana.version` (#518)
The `/search` query param was renamed from `kibana` to `kibana.version`. There are two reasons for this. The first one is to align the naming inside requirements in the manifest with the actual query param. The second is to make sure in case we have other criterias related to kibana in the future, there is space for this.

As Kibana does not use this query today, this should not affect any parts yet.

More changes:
* Cleaned up some duplicated if clauses spotted during an other PR review