Remove support for Elasticsearch requirement (#516)
So far I have not seen a package that would need to specify requirements for an Elasticsearch version as the Elasticsearch and Kibana version are tied together in the minor. Because of this removing support for Elasticsearch requirements for now. We can readd it later if we need.

This also cleans up the existing packages to remove the mention of Elasticsearch and some other examples that were still around.

As a follow up, integrations packages should be updated to get rid of the requirement. But more changes will happen to the requirements after https://github.com/elastic/package-registry/pull/515 is merged so perhaps we rather wait with this a bit more.

This change should not break any existing packages.