Discuss: Remove public dir support
The public dir was needed in the past to build packages. But with the recent changes to the registry this has become obsolete. Instead the registry can be directly pointed at a directory with packages. In the package storage this directory is called `packages`, I suggest we use this as a default. The registry itself could have its packages in the same directory.

This PR is not complete, it is more to discuss the change.