Generate package index.json files for testing
We got rid of the generated files so we are able to get rid of the public directory and the build step. But on the testing side this removed one useful tool to show changes in our generated output. To get this back, I added a test that creates all index.json files for the testdata packages.

Doing this change I found also some generated files we don't use anymore. After this is merged, I want to clean up a bit on where we put our golden files and make sure we clean it up.