Add support for owner filed in package manifest (#536)
In some cases it is important to know who the owner of a package is so this owner can be pinged for questions. In https://github.com/elastic/package-registry/issues/451 the different options around owner, authors, contributors were discussed. My take away from this discussions which lead to this PR:

* There are multiple authors over time
* Authors and contributors are mostly exchangeable terms
* There can only be 1 owner, even if it is a team.

This PR currently only implements `owner.github` because it is all we need. Later on we can add an array of `contributors` and addition owner fields if needed.

Closing https://github.com/elastic/package-registry/issues/451