Merge test and example packages (#541)
In the past there was a difference between the example and the test packages. But now as there is no build step needed anymore and all the testdata packages can also be useful to test the UI, the directory dev/package/example and testdata/package are merged. Now there is only one place in the registry with packages.

More cleanup is going to happen with where the generated data should be stored.