Proposal: Introduce default path for packages
In the past, we had public as our default path for the build packages and things just worked out of the box. With removing the build step we got rid of public and introduced support for multiple package paths. In the package-storage we used `packages` as the directory name. I am proposing to use this as our recommended default. This means, if someone has a packages directory, the user can get started without having create a config file first. It also allows us to remove the legacy public directory.

Now more catch in this PR: I renamed the config to packages_paths from package_paths. This was the original name and I suggested to change it :facepalm: But now with having packages as the default this name probably makes more sense.

If we go forward with this, we need to update the different places where the public dir and package_paths is currently used.