Cache build .tar.gz file
This introduced a cache for the tar.gz files. The code is not finished yet and hacky but it works. This also helps with https://github.com/elastic/package-registry/issues/535

Few additional notest to the code. Having the .tar.gz cached also means they could be prebuilt. This could be used during a "warmup" of the registry or in case an external storage is mounted in, the prebuilt .tar.gz could be used instead. Also these are cached across restarts.