Introduce v1 API version
This introduces a v1 prefix for all the API calls. The only calls that are not prefixed by v1 are the calls to download the .tar.gz files. As packages should not change, also path to packages should not change.

This change also makes it obvious to me that we have the prefix defined in too many places. To simplify this change, I do some previous cleanup PRs.