Fix that first package wins
In the search endpoint, follow up packages with the same name and version did overwrite the existing package. But we defined in the past that the first package found should win. This does a fix in the search endpoint and adds tests for it.

Long term I think the fix should not be in search but getPackages should already return a unique list, but this one here was easier to fix.

Also tests with multiple package paths were added so we see in case we break it in the future.