Prepare v0.6.0 release (#586)
It is time for a release. Changes to the registry should now slow down.

As soon as this PR is merged, I will tag the v0.6.0 release and open a PR to set the version to v0.7.0. All future work then goes into v0.7.0. I think that is a mistake we made here that we did not update the version number after we tagged v0.4.0