Add cache time for / index request (#631)
The current request to / was cached the the as all the static artifacts. But this info changes every time the registry is deployed with a new version. To have the new version available as soon as possible and not potentially cached for a long time, an additional configuration cache option for this endpoint was introduced.