Rename image src to path and have src as the original value from the … (#629)
All the configs specified in the manifest file by the user should be exposed without changes in our JSON output. In the case of images, the src was overwritten by the path variable. This means what was in the manifest and outputted by JSON was not a 1:1 match.

It is ok to add additional fields in the JSON output but the original fields should never be touched.

This is a breaking change and need adjustements in Kibana.