Add query to filter on asset type
This current implementation checks if any of the assets paths contains the asset.

As an example, `http://localhost:8080/search?experimental=true&asset=foo/` returns the package default_pipeline. It contains the following assets:

```
  "assets": [
    "/package/default_pipeline/0.0.2/manifest.yml",
    "/package/default_pipeline/0.0.2/docs/README.md",
    "/package/default_pipeline/0.0.2/dataset/foo/manifest.yml",
    "/package/default_pipeline/0.0.2/dataset/foo/fields/base-fields.yml",
    "/package/default_pipeline/0.0.2/dataset/foo/agent/stream/stream.yml.hbs",
    "/package/default_pipeline/0.0.2/dataset/foo/elasticsearch/ingest_pipeline/default.json"
  ],
```

This might not be expected as `foo` here is the dataset name. Probably some additional flattening of the assets is needed.