Add SQL storage indexer (beta/technical preview) (#1313)
Added new SQL storage indexer to store all the package information in a
SQL database (currently SQLite) instead of memory. The Indexer interface
has been updated to include a Close method, so database connections can
be closed at the end. This new indexer makes use of two databases so
updates can be performed faster. Updating the unused and empty database
before doing the actual swap in the service.
Moreover, it has been added a cache LRU for the search endpoint.
Search responses are one of the most expensive requests for the EPR
service, specially when there are several packages in the response. Even
if there is a CDN in front of the service, this could help to improve
performance.
During this process, code that was private in public modules has been
moved to internal modules too.

---------

Co-authored-by: Jaime Soriano Pastor <jaime.soriano@elastic.co>