Improvements to the SQL storage indexer for observability and reliability (#1334)
Added some improvements into the SQL storage indexer. All spans related to
updating database are now nested under the same parent for better visibility.
Cleanup database process just runs if cursor has been updated in that iteration.
In case of failure dropping the database, it has been added a new column in each
row of the database to set the cursor where that package belongs. All the
entrypoints whose response is a JSON (root, search, categories and package
endpoints) are unified and they use the same method. And it is allowed to
customize the default batch sizes via environment variables.