[SQL Storage indexer] Avoid storing in cache requests with package query parameter (#1378)
This PR avoids storing in the search cache requests that contain
the "package=<package>" query parameter.
Due to the potential for a large volume of unique requests, the cache
could be easily filled just with those requests or causing evictions.
Moreover, these requests (just querying for a package) typically have
rapid response times, which means not using the cache is acceptable.