[SQL indexer] Get latest packages from database (#1445)
Added a new function to return just the latest packages for those
queries that all=true has not been set. In order to get those latest
packages, package are ordered using the semantic version (major,
minor and patch numbers). This will help us to decrease the memory
usage for each query.
This new SQL query is based on the window functions:
https://sqlite.org/windowfunctions.html




* Query to retrieve just latest packages

* Fix unit tests

* Cherry-pick PR sql functions

* Update search indices paths for benchmarks

* Check release experimental if prerelease set to false in where clause

* Add missing variable when doing Scan

* Refactor in LatestFunc and AllFunc functions

* Remove unnecesary columns when doing select queries

* Add changelog entry

* Add comment

* Add a new query into benchmark

* Remove reading formatVersionMajorMinor field

* Add example of SQL query

* Added a new function to filter packages

* Refactor methods to run a query

* Rename versionBuild to be versionPrerelease

* Remove err variable to be defined in-place

* Move application level filtering out of the read lock