Fix creation of zip packages in Windows (#1490)
Package Registry was using the OS-native relative paths directly when
building zip packages. So on windows this was using back slashes, but
names in zip headers are expected to use forward slashes.

Tests were passing only because they only check the headers, and the
names there were converted to the expected format when reading the zip.